Move back to Ubuntu 14.04
This is because the license server for Siemens Geolus requires lsb-core.
That package is removed in 16.04 and I don't want to bother with getting
that working on 16.04, so I'll just drop to 14.04 and off we go. This
means a slight change in the jdk we install because we don't have a
default-jdk on 14.04 (that I know of)